- Updated TODO - Adds attribute to track Pango::Layout to Pango::Cairo. See below. - Renames Pango::Cairo.create_context to .create_pc_context, to help relieve some ambiguity. See below. - Add accessors .pango_context and .cairo_context to Pango::Cairo - Pango::Cairo.get_shape_renderer now keeps (but does not return) the data pointer. - Pango::Cairo.set_shape_renderer and .get_shape_renderer use the LAYOUT context (a Pango::Context) as opposed to the Pango::Cairo context. This may evolve further, in the future. - t/04-shape.t is now executing its shape renderer function, correctly. However there is still NO CHANGE in output. However any progress, is good progress
+ COMMENT: It remains to be seen how Pango::Cairo is INTENDED to be
used. The changes made in this commit attempt to correct much of the
ambiguity, but there it remains. Between Pango::Context, Pango::Cairo
context and Cairo context, there are just too many contexts, and it is
still not set in stone which one is the right one. I can only go by the
example code I can find on the Internet.

It's quite possible that all of this is WRONG and that a shape renderer
can be used on ANY context. To anyone who reads this: Clarification
would be greatly appreciated.